+++ adapters/AppLovin/CHANGELOG.md
@@ -4,6 +4,7 @@
 - Replaced the deprecated AppLovin SDK's initialization APIs with its new APIs.
 - Prevented multiple initializations of AppLovin SDK when multiple AppLovin SDK keys are provided.
 - Replaced the deprecated AppLovin SDK's bid token collection API with its new one.
+- Deprecated the `muteAudio` property, use `GADMobileAds.sharedInstance.applicationMuted` instead.
 
 #### [Version 13.0.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/applovin/AppLovinAdapter-13.0.1.0.zip)
 - Verified compatibility with AppLovin SDK 13.0.1.